python/smolagents-basic/pyproject.toml
[project]
name = "smolagents-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "flask>=3.1.2",
    "google-api-python-client>=2.182.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "litellm>=1.77.1",
    "python-dotenv>=1.1.1",
    "smolagents>=1.21.3",
    "torch>=2.8.0",
    "transformers>=4.56.1",
]
